商品返回接口

+++ YYG/src/com/yyg/model/UserLotteryMappingTable.java
@@ -20,6 +20,6 @@ public class UserLotteryMappingTable {
 	public Order order; 
 	
 	@DatabaseField
-	public long joinTime;
+	public int joinTime;
 	
 }
